Update versionCode to 3 and versionName to 1.2. Refactor to a.
