--- app/src/androidTest/java/ms/appcenter/sampleapp/android/ExampleInstrumentedTest.java
@@ -1,40 +1,0 @@
-package ms.appcenter.sampleapp.android;
-
-import android.content.Context;
-
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-import androidx.test.platform.app.InstrumentationRegistry;
-
-import org.junit.Rule;
-import org.junit.After;
-import com.microsoft.appcenter.espresso.Factory;
-import com.microsoft.appcenter.espresso.ReportHelper;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Instrumentation test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
-    @Rule
-    public ReportHelper reportHelper = Factory.getReportHelper();
-
-    @After
-    public void TearDown() {
-        reportHelper.label("Stopping App");
-    }
-
-    @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-
-        assertEquals("ms.appcenter.sampleapp.android", appContext.getPackageName());
-    }
-}
