+++ app/src/main/java/ms/appcenter/sampleapp/a/AnalyticsActivity.java
@@ -1,4 +1,4 @@
-package ms.appcenter.sampleapp.android;
+package ms.appcenter.sampleapp.a;
 
 import android.app.Dialog;
 import android.os.Bundle;
@@ -10,6 +10,7 @@ import android.widget.Button;
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
+import ms.appcenter.sampleapp.android.R;
 
 import com.microsoft.appcenter.analytics.Analytics;
