+++ app/src/main/java/ms/appcenter/sampleapp/a/BuildActivity.java
@@ -1,4 +1,4 @@
-package ms.appcenter.sampleapp.android;
+package ms.appcenter.sampleapp.a;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,15 +6,15 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
+import ms.appcenter.sampleapp.android.R;
 
-public class DistributeActivity extends Fragment {
+public class BuildActivity extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.distribute_root, container, false);
+                R.layout.build_root, container, false);
         return rootView;
     }
 }
-
