+++ app/src/main/java/ms/appcenter/sampleapp/a/WelcomeActivity.java
@@ -1,4 +1,4 @@
-package ms.appcenter.sampleapp.android;
+package ms.appcenter.sampleapp.a;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,13 +6,14 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
+import ms.appcenter.sampleapp.android.R;
 
-public class PushActivity extends Fragment {
+public class WelcomeActivity extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.push_root, container, false);
+                R.layout.welcome_root, container, false);
         return rootView;
     }
 }
