+++ app/src/main/res/layout/activity_root.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="ms.appcenter.sampleapp.android.MainActivity">
+    tools:context="ms.appcenter.sampleapp.a.MainActivity">
 
     <androidx.viewpager.widget.ViewPager
         android:id="@+id/container"
